Module work
This is a major study I've done over the passed few days. It includes examples of how to use some of the different modules in Python. This is no way an exhaustive list. This is to build an understanding of the more fun side of Python.

You will module code including, but not limited to:

os
pandas
tkinter
matplotlib
numpy
turtle